+++ mzmine-community/src/main/java/io/github/mzmine/datamodel/features/FeatureListRow.java
@@ -147,7 +147,7 @@ public interface FeatureListRow extends ModularDataModel {
    * Returns average RI for features on this row
    */
 
-  Integer getAverageRI();
+  Float getAverageRI();
 
   /**
    * Sets average rt for this row
